requirements.txt
loguru==0.7.2
numpy==1.26.4
Pillow==10.3.0
